Manage transaction list in hook
